This time, a socket was created to connect to the real server and the target website. Standard HTTP request first line and Host header were generated. Subsequently, the webpage content is received from the origin server, and the received content is fully forwarded to the client browser. Finally, the data returned by the server is saved to the local disk.
This time, a socket was created to connect to the real server and the target website. Standard HTTP request first line and Host header were generated. Subsequently, the webpage content is received from the origin server, and the received content is fully forwarded to the client browser. Finally, the data returned by the server is saved to the local disk.